+++ packages/theme/plugins/buttonGroup.mjs
@@ -20,286 +20,286 @@ export function buttonGroup({ addComponents }) {
   };
 
   const variantsMap = {
-    ".buttonGroup-default-solid": {
+    ".button-group-default-solid": {
       "@apply bg-[#1f2937]": {},
     },
-    ".buttonGroup-default-solid:hover": {
+    ".button-group-default-solid:hover": {
       "@apply bg-[#1f2937]": {},
     },
-    ".buttonGroup-default-solid:disabled": {
+    ".button-group-default-solid:disabled": {
       "@apply bg-[rgba(31,41,55,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-default-outline": {
+    ".button-group-default-outline": {
       "@apply bg-[#1f2937] border border-[#1f2937]": {},
     },
-    ".buttonGroup-default-outline:hover": {
+    ".button-group-default-outline:hover": {
       "@apply bg-[#1f2937] border border-[#1f2937]": {},
     },
-    ".buttonGroup-default-outline:disabled": {
+    ".button-group-default-outline:disabled": {
       "@apply bg-[rgba(31,41,55,0.5)] border border-[rgba(31,41,55,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-default-outline-dashed": {
+    ".button-group-default-outline-dashed": {
       "@apply bg-[#1f2937] border border-dashed border-[#1f2937]": {},
     },
-    ".buttonGroup-default-outline-dashed:hover": {
+    ".button-group-default-outline-dashed:hover": {
       "@apply bg-[#1f2937] border border-dashed border-[#1f2937]": {},
     },
-    ".buttonGroup-default-outline-dashed:disabled": {
+    ".button-group-default-outline-dashed:disabled": {
       "@apply bg-[rgba(31,41,55,0.5)] border border-dashed border-[rgba(31,41,55,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-default-no-outline": {
+    ".button-group-default-no-outline": {
       "@apply bg-[#1f2937]": {},
     },
-    ".buttonGroup-default-no-outline:hover": {
+    ".button-group-default-no-outline:hover": {
       "@apply bg-[#1f2937]": {},
     },
-    ".buttonGroup-default-no-outline:disabled": {
+    ".button-group-default-no-outline:disabled": {
       "@apply bg-[rgba(31,41,55,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-default-glowing": {
+    ".button-group-default-glowing": {
       "@apply bg-[#1f2937] shadow-[0_0_6px_#1f293780]": {},
     },
-    ".buttonGroup-default-glowing:hover": {
+    ".button-group-default-glowing:hover": {
       "@apply bg-[#111827] shadow-[0_0_8px_#11182780]": {},
     },
-    ".buttonGroup-default-glowing:disabled": {
+    ".button-group-default-glowing:disabled": {
       "@apply bg-[rgba(31,41,55,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-primary-solid": {
+    ".button-group-primary-solid": {
       "@apply bg-[#2563eb]": {},
     },
-    ".buttonGroup-primary-solid:hover": {
+    ".button-group-primary-solid:hover": {
       "@apply bg-[#2563eb]": {},
     },
-    ".buttonGroup-primary-solid:disabled": {
+    ".button-group-primary-solid:disabled": {
       "@apply bg-[rgba(37,99,235,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-primary-outline": {
+    ".button-group-primary-outline": {
       "@apply bg-[#2563eb] border border-[#2563eb]": {},
     },
-    ".buttonGroup-primary-outline:hover": {
+    ".button-group-primary-outline:hover": {
       "@apply bg-[#2563eb] border border-[#2563eb]": {},
     },
-    ".buttonGroup-primary-outline:disabled": {
+    ".button-group-primary-outline:disabled": {
       "@apply bg-[rgba(37,99,235,0.5)] border border-[rgba(37,99,235,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-primary-outline-dashed": {
+    ".button-group-primary-outline-dashed": {
       "@apply bg-[#2563eb] border border-dashed border-[#2563eb]": {},
     },
-    ".buttonGroup-primary-outline-dashed:hover": {
+    ".button-group-primary-outline-dashed:hover": {
       "@apply bg-[#2563eb] border border-dashed border-[#2563eb]": {},
     },
-    ".buttonGroup-primary-outline-dashed:disabled": {
+    ".button-group-primary-outline-dashed:disabled": {
       "@apply bg-[rgba(37,99,235,0.5)] border border-dashed border-[rgba(37,99,235,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-primary-no-outline": {
+    ".button-group-primary-no-outline": {
       "@apply bg-[#2563eb]": {},
     },
-    ".buttonGroup-primary-no-outline:hover": {
+    ".button-group-primary-no-outline:hover": {
       "@apply bg-[#2563eb]": {},
     },
-    ".buttonGroup-primary-no-outline:disabled": {
+    ".button-group-primary-no-outline:disabled": {
       "@apply bg-[rgba(37,99,235,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-primary-glowing": {
+    ".button-group-primary-glowing": {
       "@apply bg-[#2563eb] shadow-[0_0_6px_#2563eb80]": {},
     },
-    ".buttonGroup-primary-glowing:hover": {
+    ".button-group-primary-glowing:hover": {
       "@apply bg-[#1d4ed8] shadow-[0_0_8px_#1d4ed880]": {},
     },
-    ".buttonGroup-primary-glowing:disabled": {
+    ".button-group-primary-glowing:disabled": {
       "@apply bg-[rgba(37,99,235,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-secondary-solid": {
+    ".button-group-secondary-solid": {
       "@apply bg-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-solid:hover": {
+    ".button-group-secondary-solid:hover": {
       "@apply bg-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-solid:disabled": {
+    ".button-group-secondary-solid:disabled": {
       "@apply bg-[rgba(79,70,229,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-secondary-outline": {
+    ".button-group-secondary-outline": {
       "@apply bg-[#4f46e5] border border-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-outline:hover": {
+    ".button-group-secondary-outline:hover": {
       "@apply bg-[#4f46e5] border border-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-outline:disabled": {
+    ".button-group-secondary-outline:disabled": {
       "@apply bg-[rgba(79,70,229,0.5)] border border-[rgba(79,70,229,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-secondary-outline-dashed": {
+    ".button-group-secondary-outline-dashed": {
       "@apply bg-[#4f46e5] border border-dashed border-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-outline-dashed:hover": {
+    ".button-group-secondary-outline-dashed:hover": {
       "@apply bg-[#4f46e5] border border-dashed border-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-outline-dashed:disabled": {
+    ".button-group-secondary-outline-dashed:disabled": {
       "@apply bg-[rgba(79,70,229,0.5)] border border-dashed border-[rgba(79,70,229,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-secondary-no-outline": {
+    ".button-group-secondary-no-outline": {
       "@apply bg-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-no-outline:hover": {
+    ".button-group-secondary-no-outline:hover": {
       "@apply bg-[#4f46e5]": {},
     },
-    ".buttonGroup-secondary-no-outline:disabled": {
+    ".button-group-secondary-no-outline:disabled": {
       "@apply bg-[rgba(79,70,229,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-secondary-glowing": {
+    ".button-group-secondary-glowing": {
       "@apply bg-[#4f46e5] shadow-[0_0_6px_#4f46e580]": {},
     },
-    ".buttonGroup-secondary-glowing:hover": {
+    ".button-group-secondary-glowing:hover": {
       "@apply bg-[#4338ca] shadow-[0_0_8px_#4338ca80]": {},
     },
-    ".buttonGroup-secondary-glowing:disabled": {
+    ".button-group-secondary-glowing:disabled": {
       "@apply bg-[rgba(79,70,229,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-success-solid": {
+    ".button-group-success-solid": {
       "@apply bg-[#16a34a]": {},
     },
-    ".buttonGroup-success-solid:hover": {
+    ".button-group-success-solid:hover": {
       "@apply bg-[#16a34a]": {},
     },
-    ".buttonGroup-success-solid:disabled": {
+    ".button-group-success-solid:disabled": {
       "@apply bg-[rgba(22,163,74,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-success-outline": {
+    ".button-group-success-outline": {
       "@apply bg-[#16a34a] border border-[#16a34a]": {},
     },
-    ".buttonGroup-success-outline:hover": {
+    ".button-group-success-outline:hover": {
       "@apply bg-[#16a34a] border border-[#16a34a]": {},
     },
-    ".buttonGroup-success-outline:disabled": {
+    ".button-group-success-outline:disabled": {
       "@apply bg-[rgba(22,163,74,0.5)] border border-[rgba(22,163,74,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-success-outline-dashed": {
+    ".button-group-success-outline-dashed": {
       "@apply bg-[#16a34a] border border-dashed border-[#16a34a]": {},
     },
-    ".buttonGroup-success-outline-dashed:hover": {
+    ".button-group-success-outline-dashed:hover": {
       "@apply bg-[#16a34a] border border-dashed border-[#16a34a]": {},
     },
-    ".buttonGroup-success-outline-dashed:disabled": {
+    ".button-group-success-outline-dashed:disabled": {
       "@apply bg-[rgba(22,163,74,0.5)] border border-dashed border-[rgba(22,163,74,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-success-no-outline": {
+    ".button-group-success-no-outline": {
       "@apply bg-[#16a34a]": {},
     },
-    ".buttonGroup-success-no-outline:hover": {
+    ".button-group-success-no-outline:hover": {
       "@apply bg-[#16a34a]": {},
     },
-    ".buttonGroup-success-no-outline:disabled": {
+    ".button-group-success-no-outline:disabled": {
       "@apply bg-[rgba(22,163,74,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-success-glowing": {
+    ".button-group-success-glowing": {
       "@apply bg-[#16a34a] shadow-[0_0_6px_#16a34a80]": {},
     },
-    ".buttonGroup-success-glowing:hover": {
+    ".button-group-success-glowing:hover": {
       "@apply bg-[#15803d] shadow-[0_0_8px_#15803d80]": {},
     },
-    ".buttonGroup-success-glowing:disabled": {
+    ".button-group-success-glowing:disabled": {
       "@apply bg-[rgba(22,163,74,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-warning-solid": {
+    ".button-group-warning-solid": {
       "@apply bg-[#d97706]": {},
     },
-    ".buttonGroup-warning-solid:hover": {
+    ".button-group-warning-solid:hover": {
       "@apply bg-[#d97706]": {},
     },
-    ".buttonGroup-warning-solid:disabled": {
+    ".button-group-warning-solid:disabled": {
       "@apply bg-[rgba(217,119,6,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-warning-outline": {
+    ".button-group-warning-outline": {
       "@apply bg-[#d97706] border border-[#d97706]": {},
     },
-    ".buttonGroup-warning-outline:hover": {
+    ".button-group-warning-outline:hover": {
       "@apply bg-[#d97706] border border-[#d97706]": {},
     },
-    ".buttonGroup-warning-outline:disabled": {
+    ".button-group-warning-outline:disabled": {
       "@apply bg-[rgba(217,119,6,0.5)] border border-[rgba(217,119,6,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-warning-outline-dashed": {
+    ".button-group-warning-outline-dashed": {
       "@apply bg-[#d97706] border border-dashed border-[#d97706]": {},
     },
-    ".buttonGroup-warning-outline-dashed:hover": {
+    ".button-group-warning-outline-dashed:hover": {
       "@apply bg-[#d97706] border border-dashed border-[#d97706]": {},
     },
-    ".buttonGroup-warning-outline-dashed:disabled": {
+    ".button-group-warning-outline-dashed:disabled": {
       "@apply bg-[rgba(217,119,6,0.5)] border border-dashed border-[rgba(217,119,6,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-warning-no-outline": {
+    ".button-group-warning-no-outline": {
       "@apply bg-[#d97706]": {},
     },
-    ".buttonGroup-warning-no-outline:hover": {
+    ".button-group-warning-no-outline:hover": {
       "@apply bg-[#d97706]": {},
     },
-    ".buttonGroup-warning-no-outline:disabled": {
+    ".button-group-warning-no-outline:disabled": {
       "@apply bg-[rgba(217,119,6,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-warning-glowing": {
+    ".button-group-warning-glowing": {
       "@apply bg-[#d97706] shadow-[0_0_6px_#d9770680]": {},
     },
-    ".buttonGroup-warning-glowing:hover": {
+    ".button-group-warning-glowing:hover": {
       "@apply bg-[#b45309] shadow-[0_0_8px_#b4530980]": {},
     },
-    ".buttonGroup-warning-glowing:disabled": {
+    ".button-group-warning-glowing:disabled": {
       "@apply bg-[rgba(217,119,6,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-error-solid": {
+    ".button-group-error-solid": {
       "@apply bg-[#dc2626]": {},
     },
-    ".buttonGroup-error-solid:hover": {
+    ".button-group-error-solid:hover": {
       "@apply bg-[#dc2626]": {},
     },
-    ".buttonGroup-error-solid:disabled": {
+    ".button-group-error-solid:disabled": {
       "@apply bg-[rgba(220,38,38,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-error-outline": {
+    ".button-group-error-outline": {
       "@apply bg-[#dc2626] border border-[#dc2626]": {},
     },
-    ".buttonGroup-error-outline:hover": {
+    ".button-group-error-outline:hover": {
       "@apply bg-[#dc2626] border border-[#dc2626]": {},
     },
-    ".buttonGroup-error-outline:disabled": {
+    ".button-group-error-outline:disabled": {
       "@apply bg-[rgba(220,38,38,0.5)] border border-[rgba(220,38,38,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-error-outline-dashed": {
+    ".button-group-error-outline-dashed": {
       "@apply bg-[#dc2626] border border-dashed border-[#dc2626]": {},
     },
-    ".buttonGroup-error-outline-dashed:hover": {
+    ".button-group-error-outline-dashed:hover": {
       "@apply bg-[#dc2626] border border-dashed border-[#dc2626]": {},
     },
-    ".buttonGroup-error-outline-dashed:disabled": {
+    ".button-group-error-outline-dashed:disabled": {
       "@apply bg-[rgba(220,38,38,0.5)] border border-dashed border-[rgba(220,38,38,0.5)] opacity-50 cursor-not-allowed":
         {},
     },
-    ".buttonGroup-error-no-outline": {
+    ".button-group-error-no-outline": {
       "@apply bg-[#dc2626]": {},
     },
-    ".buttonGroup-error-no-outline:hover": {
+    ".button-group-error-no-outline:hover": {
       "@apply bg-[#dc2626]": {},
     },
-    ".buttonGroup-error-no-outline:disabled": {
+    ".button-group-error-no-outline:disabled": {
       "@apply bg-[rgba(220,38,38,0.5)] opacity-50 cursor-not-allowed": {},
     },
-    ".buttonGroup-error-glowing": {
+    ".button-group-error-glowing": {
       "@apply bg-[#dc2626] shadow-[0_0_6px_#dc262680]": {},
     },
-    ".buttonGroup-error-glowing:hover": {
+    ".button-group-error-glowing:hover": {
       "@apply bg-[#b91c1c] shadow-[0_0_8px_#b91c1c80]": {},
     },
-    ".buttonGroup-error-glowing:disabled": {
+    ".button-group-error-glowing:disabled": {
       "@apply bg-[rgba(220,38,38,0.5)] opacity-50 cursor-not-allowed": {},
     },
   };