+++ packages/jest-config/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @aristobyte-ui/jest-config
 
+## 1.0.113
+
 ## 1.0.112
 
 ## 1.0.111