+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @aristobyte-ui/theme
 
+## 1.0.113
+
+### Patch Changes
+
+- feat: make the packages both remotely and dev accessible
+
+- Updated dependencies []:
+  - @aristobyte-ui/eslint-config@1.0.113
+
 ## 1.0.112
 
 ### Patch Changes